refactor: migrate browsingContext methods to frame
Also improves docs for navigation.
