+++ docs/api/index.md
@@ -94,6 +94,7 @@ sidebar_label: API
 | [FrameEvents](./puppeteer.frameevents.md)                                   |                                                                                                                                                                                     |
 | [FrameWaitForFunctionOptions](./puppeteer.framewaitforfunctionoptions.md)   |                                                                                                                                                                                     |
 | [GeolocationOptions](./puppeteer.geolocationoptions.md)                     |                                                                                                                                                                                     |
+| [GoToOptions](./puppeteer.gotooptions.md)                                   |                                                                                                                                                                                     |
 | [InterceptResolutionState](./puppeteer.interceptresolutionstate.md)         |                                                                                                                                                                                     |
 | [InternalNetworkConditions](./puppeteer.internalnetworkconditions.md)       |                                                                                                                                                                                     |
 | [JSCoverageEntry](./puppeteer.jscoverageentry.md)                           | The CoverageEntry class for JavaScript                                                                                                                                              |
