+++ docs/api/puppeteer.frame.md
@@ -85,7 +85,7 @@ console.log(text);
 | [evaluate(pageFunction, args)](./puppeteer.frame.evaluate.md)                        |           | Behaves identically to [Page.evaluate()](./puppeteer.page.evaluate.md) except it's run within the the context of this frame.                                                                                                                                               |
 | [evaluateHandle(pageFunction, args)](./puppeteer.frame.evaluatehandle.md)            |           | Behaves identically to [Page.evaluateHandle()](./puppeteer.page.evaluatehandle.md) except it's run within the context of this frame.                                                                                                                                       |
 | [focus(selector)](./puppeteer.frame.focus.md)                                        |           | Focuses the first element that matches the <code>selector</code>.                                                                                                                                                                                                          |
-| [goto(url, options)](./puppeteer.frame.goto.md)                                      |           | Navigates a frame to the given url.                                                                                                                                                                                                                                        |
+| [goto(url, options)](./puppeteer.frame.goto.md)                                      |           | Navigates the frame to the given <code>url</code>.                                                                                                                                                                                                                         |
 | [hover(selector)](./puppeteer.frame.hover.md)                                        |           | Hovers the pointer over the center of the first element that matches the <code>selector</code>.                                                                                                                                                                            |
 | [isDetached()](./puppeteer.frame.isdetached.md)                                      |           | Is<code>true</code> if the frame has been detached. Otherwise, <code>false</code>.                                                                                                                                                                                         |
 | [isOOPFrame()](./puppeteer.frame.isoopframe.md)                                      |           | Is <code>true</code> if the frame is an out-of-process (OOP) frame. Otherwise, <code>false</code>.                                                                                                                                                                         |
