+++ packages/puppeteer-core/src/cdp/LifecycleWatcher.ts
@@ -27,13 +27,28 @@ import {type CdpFrame} from './Frame.js';
 import {FrameManagerEvent} from './FrameManager.js';
 import {type CdpHTTPRequest} from './HTTPRequest.js';
 import {type NetworkManager, NetworkManagerEvent} from './NetworkManager.js';
+
 /**
  * @public
  */
 export type PuppeteerLifeCycleEvent =
+  /**
+   * Waits for the 'load' event.
+   */
   | 'load'
+  /**
+   * Waits for the 'DOMContentLoaded' event.
+   */
   | 'domcontentloaded'
+  /**
+   * Waits till there are no more than 0 network connections for at least `500`
+   * ms.
+   */
   | 'networkidle0'
+  /**
+   * Waits till there are no more than 2 network connections for at least `500`
+   * ms.
+   */
   | 'networkidle2';
 
 /**
